refactor: mv to middleware.py
