+++ iaso/middleware.py
@@ -1,5 +1,8 @@
+from collections import defaultdict
+
 from django.urls import Resolver404, resolve
 from djangorestframework_camel_case.middleware import CamelCaseMiddleWare
+from querycount.middleware import QueryCountMiddleware
 
 
 class CustomCamelCaseMiddleWare(CamelCaseMiddleWare):
@@ -14,3 +17,23 @@ class CustomCamelCaseMiddleWare(CamelCaseMiddleWare):
         if route.startswith(("/api/v2/profiles", "/api/validation-workflows")):
             return super().__call__(request)
         return self.get_response(request)
+
+
+class SafeQueryCountMiddleware(QueryCountMiddleware):
+    """QueryCountMiddleware that handles dynamically added database aliases.
+
+    The upstream middleware builds a fixed stats dict at __init__ time from
+    the aliases known to ``connections.all()``. When a database alias is
+    registered dynamically, ``_count_queries`` crashes with a KeyError.
+    This subclass uses ``defaultdict`` so that any alias access auto-creates
+    its stats entry.
+    """
+
+    @staticmethod
+    def _empty_stats():
+        return {"writes": 0, "reads": 0, "total": 0, "duplicates": 0}
+
+    def _reset_stats(self):
+        super()._reset_stats()
+        self.stats["request"] = defaultdict(self._empty_stats, self.stats["request"])
+        self.stats["response"] = defaultdict(self._empty_stats, self.stats["response"])
